backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic-settings==2.4.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0